fix sending short instead of long

+++ public-app/src/domain/urlResolved/UrlResolved.ts
@@ -8,7 +8,7 @@ export class UrlResolved {
 
     public static createFromResolve(resolve: Resolve): UrlResolved {
         return new UrlResolved(
-            resolve.long,
+            resolve.short,
             new Date(),
         );
     }
